ocsApiUserProvisioningManager

ocsApiUserProvisioningManager

ocsApiUserProvisioningManager

Format

R6Class object.

Value

Object of R6Class for modelling an ocsManager for Webdav API

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

ocs4R::ocs4RLogger -> ocs4R::ocsManager -> ocsApiUserProvisioningManager

Methods

Public methods

ocsApiUserProvisioningManager$new()

ocsApiUserProvisioningManager$addUser()

ocsApiUserProvisioningManager$getUsers()

ocsApiUserProvisioningManager$getUser()

ocsApiUserProvisioningManager$editUser()

ocsApiUserProvisioningManager$editUserDisplayName()

ocsApiUserProvisioningManager$editUserEmail()

ocsApiUserProvisioningManager$editUserPassword()

ocsApiUserProvisioningManager$editUserQuota()

ocsApiUserProvisioningManager$enableUser()

ocsApiUserProvisioningManager$disableUser()

ocsApiUserProvisioningManager$deleteUser()

ocsApiUserProvisioningManager$getUserGroups()

ocsApiUserProvisioningManager$addToGroup()

ocsApiUserProvisioningManager$removeFromGroup()

ocsApiUserProvisioningManager$createSubadmin()

ocsApiUserProvisioningManager$removeSubadmin()

ocsApiUserProvisioningManager$getSubadminGroups()

ocsApiUserProvisioningManager$getGroups()

ocsApiUserProvisioningManager$addGroup()

ocsApiUserProvisioningManager$getGroup()

ocsApiUserProvisioningManager$deleteGroup()

ocsApiUserProvisioningManager$getSubadmins()

ocsApiUserProvisioningManager$clone()

Inherited methods

ocs4R::ocs4RLogger$ERROR()

ocs4R::ocs4RLogger$INFO()

ocs4R::ocs4RLogger$WARN()

ocs4R::ocs4RLogger$getClass()

ocs4R::ocs4RLogger$getClassName()

ocs4R::ocs4RLogger$logger()

ocs4R::ocsManager$connect()

ocs4R::ocsManager$getAPISharingManager()

ocs4R::ocsManager$getAPIUserProvisioningManager()

ocs4R::ocsManager$getAPIWebdavManager()

ocs4R::ocsManager$getCapabilities()

ocs4R::ocsManager$getVersion()

Method new()

Initialize manager

Usage

ocsApiUserProvisioningManager$new(
  url,
  user,
  pwd,
  logger = NULL,
  keyring_backend = "env"
)

Arguments

url

url

user

user

pwd

pwd

logger

logger

keyring_backend

backend to use with keyring. Default is NULL

Method addUser()

Adds a user given a userid (required). All other fields (email, password, groups) are optional for the user creation. Returns TRUE if the user is added, FALSE otherwise.

Usage

ocsApiUserProvisioningManager$addUser(
  userid,
  email = NULL,
  password = NULL,
  groups = NULL
)

Arguments

userid

user ID

email

email

password

user password

groups

groups

Method getUsers()

Get the list of users. This method returns a vector of class 'character' giving the user IDs available in the OCS cloud plateform.

Usage

ocsApiUserProvisioningManager$getUsers()

Method getUser()

Get the user details from its userid. If the argument pretty is set to TRUE, this will return an object of class data.frame, otherwise (by default) it returns an object of class list.

Usage

ocsApiUserProvisioningManager$getUser(userid, pretty = FALSE)

Arguments

userid

user ID

pretty

pretty

Method editUser()

Edits a user, identifier by a userid. The user property to be edited should be set using its key (eg display) and the value to be modified for this key. Returns TRUE if the user is edited, FALSE otherwise.

Usage

ocsApiUserProvisioningManager$editUser(userid, key, value)

Arguments

userid

user ID

key

key

value

value

Method editUserDisplayName()

Edits a user display name.

Usage

ocsApiUserProvisioningManager$editUserDisplayName(userid, displayName)

Arguments

userid

user ID

displayName

display name

Method editUserEmail()

Edits a user email

Usage

ocsApiUserProvisioningManager$editUserEmail(userid, email)

Arguments

userid

user ID

email

email

Method editUserPassword()

Edits a user password

Usage

ocsApiUserProvisioningManager$editUserPassword(userid, password)

Arguments

userid

user ID

password

password

Method editUserQuota()

Edits a user quota

Usage

ocsApiUserProvisioningManager$editUserQuota(userid, quota)

Arguments

userid

user ID

quota

quota

Method enableUser()

Enables a user

Usage

ocsApiUserProvisioningManager$enableUser(userid)

Arguments

userid

user ID

Returns

TRUE if enabled, FALSE otherwise

Method disableUser()

Disables a user

Usage

ocsApiUserProvisioningManager$disableUser(userid)

Arguments

userid

user ID

Returns

TRUE if disabled, FALSE otherwise

Method deleteUser()

Deletes a user

Usage

ocsApiUserProvisioningManager$deleteUser(userid)

Arguments

userid

user ID

Returns

TRUE if deleted, FALSE otherwise

Method getUserGroups()

Get user groups

Usage

ocsApiUserProvisioningManager$getUserGroups(userid)

Arguments

userid

user ID

Returns

the user groups

Method addToGroup()

Adds a user to a group.

Usage

ocsApiUserProvisioningManager$addToGroup(userid, groupid)

Arguments

userid

user ID

groupid

group ID

Returns

TRUE if added, FALSE otherwise

Method removeFromGroup()

Removes a user from a group.

Usage

ocsApiUserProvisioningManager$removeFromGroup(userid, groupid)

Arguments

userid

user ID

groupid

group ID

Returns

TRUE if removed, FALSE otherwise

Method createSubadmin()

Creates a subadmin

Usage

ocsApiUserProvisioningManager$createSubadmin()

Method removeSubadmin()

Removes a subadmin

Usage

ocsApiUserProvisioningManager$removeSubadmin()

Method getSubadminGroups()

Get subadmin groups

Usage

ocsApiUserProvisioningManager$getSubadminGroups()

Method getGroups()

Get the list of groups. This method returns a vector of class 'character' giving the usergroups IDs

Usage

ocsApiUserProvisioningManager$getGroups(
  search = NULL,
  limit = NULL,
  offset = NULL
)

Arguments

search

search

limit

limit

offset

offset

Method addGroup()

Adds a group

Usage

ocsApiUserProvisioningManager$addGroup(groupid)

Arguments

groupid

group ID

Returns

TRUE if added, FALSE

Method getGroup()

Gets a group

Usage

ocsApiUserProvisioningManager$getGroup(groupid)

Arguments

groupid

group ID

Returns

the group as list including the group ID and the list of users

Method deleteGroup()

Deletes a group

Usage

ocsApiUserProvisioningManager$deleteGroup(groupid)

Arguments

groupid

group ID

Returns

TRUE if deleted, FALSE

Method getSubadmins()

Get subadmins

Usage

ocsApiUserProvisioningManager$getSubadmins()

Method clone()

The objects of this class are cloneable with this method.

Usage

ocsApiUserProvisioningManager$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.