ocsApiWebdavManager

ocsApiWebdavManager

Format

R6Class object.

Value

Object of R6Class for modelling an ocsManager for Webdav API

General Methods (inherited from 'ocsManager')

new(url, user, pwd, logger, keyring_backend)

This method is used to instantiate an ocsApiWebdavManager. The user/pwd are mandatory in order to connect to 'ocs'.

The logger can be either NULL, "INFO" (with minimum logs), or "DEBUG" (for complete curl http calls logs).

The keyring_backend can be set to use a different backend for storing the user password with keyring (Default value is 'env').

connect()

A method to connect to 'ocs' and set version/capabilities

getVersion()

Get the 'ocs' server version

getCapabilities()

Get the 'ocs' server capabilities

WebDAV methods

getWebdavRoot()

Get the 'ocs' WebDAV root URL

listFiles(relPath)

WebDAV method to list folders/files given a relative path. The relative path is set to "/" by default, which corresponds to the root of the 'ocs' repository.

makeCollection(name, relPath)

WebDAV method to make a collection. By default relPath is set to "/" (root). The name is the name of the new collection to be created. The function is recursive in the sense that a name can be provided as relative path of a collection tree (eg newfolder1/newfolder2/newfolder3), the function will create recursively as many collections are handled in the name.

uploadFile(filename, relPath, delete_if_existing)

WebDAV method to upload a file. By default relPath is set to "/" (root).

deleteFile(filename, relPath)

WebDAV method to delete a file. By default relPath is set to "/" (root).

getPublicFile(share_token)

Get details of a shared public file given its share token

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>