ocsManager

ocsManager

ocsManager

Format

R6Class object.

Value

Object of R6Class for modelling an ocsManager

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super class

ocs4R::ocs4RLogger -> ocsManager

Public fields

apis

list of APIs

Methods

Public methods

ocsManager$new()

ocsManager$connect()

ocsManager$getVersion()

ocsManager$getCapabilities()

ocsManager$getAPIWebdavManager()

ocsManager$getAPISharingManager()

ocsManager$getAPIUserProvisioningManager()

ocsManager$clone()

Inherited methods

ocs4R::ocs4RLogger$ERROR()

ocs4R::ocs4RLogger$INFO()

ocs4R::ocs4RLogger$WARN()

ocs4R::ocs4RLogger$getClass()

ocs4R::ocs4RLogger$getClassName()

ocs4R::ocs4RLogger$logger()

Method new()

This method is used to instantiate an ocsManager. The user/pwd are mandatory in order to connect to 'ocs'. The logger can be either NULL, "INFO" (with minimum logs), or "DEBUG" (for complete curl http calls logs).

The keyring_backend can be set to use a different backend for storing the user password with keyring (Default value is NULL, meaning the password is stored as private field).

Usage

ocsManager$new(url, user, pwd, logger = NULL, keyring_backend = NULL)

Arguments

url

url

user

user

pwd

pwd

logger

logger type

keyring_backend

keyring back-end. Default is NULL

Method connect()

Method to connect to 'ocs' and set version/capabilities

Usage

ocsManager$connect()

Method getVersion()

Get the 'ocs' server version

Usage

ocsManager$getVersion()

Method getCapabilities()

Get the 'ocs' server capabilities

Usage

ocsManager$getCapabilities()

Method getAPIWebdavManager()

Get the Webdav API manager

Usage

ocsManager$getAPIWebdavManager()

Returns

an instance of ocsApiWebdavManager

Method getAPISharingManager()

Get the Sharing API manager

Usage

ocsManager$getAPISharingManager()

Returns

an instance of ocsApiSharingManager

Method getAPIUserProvisioningManager()

Get the User Provisioning API manager

Usage

ocsManager$getAPIUserProvisioningManager()

Returns

an instance of ocsApiUserProvisioningManager

Method clone()

The objects of this class are cloneable with this method.

Usage

ocsManager$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \dontrun{
   #Not Run:
   #Connect to an OCS API
   OCS <- ocsManager$new(url = ocs_url, user = ocs_user, pwd = ocs_pwd, logger = "DEBUG")
#> Error: object 'ocs_url' not found
   version <- OCS$getVersion()
#> Error: object 'OCS' not found
   caps <- OCS$getCapabilities()
#> Error: object 'OCS' not found
   
   #OCS User Provisioning API
   #-------------------------------------
   #users
   users <- OCS$getUsers() #get users
#> Error: object 'OCS' not found
   user <- OCS$getUser("admin") #get a user
#> Error: object 'OCS' not found
   user.df <- OCS$getUser("admin", TRUE) #the same user as data.frame
#> Error: object 'OCS' not found
   added <- OCS$addUser("john.doe", password = "ocs4john") #add a user
#> Error: object 'OCS' not found
   disabled <- OCS$disableUser("john.doe") #disable a user
#> Error: object 'OCS' not found
   enabled <- OCS$enableUser("john.doe") #enable auser
#> Error: object 'OCS' not found
   edited <- OCS$editUser("john.doe", key = "display", value = "John Doe") #edit user
#> Error: object 'OCS' not found
   #edit some user field
   edited2 <- OCS$editUserDisplayName("john.doe", displayName = "John Doe Jr.") 
#> Error: object 'OCS' not found
   deleted <- OCS$deleteUser("john.doe")
#> Error: object 'OCS' not found
   
   #groups
   admingroups <- OCS$getUserGroups("admin")
#> Error: object 'OCS' not found
   groups <- OCS$getGroups()
#> Error: object 'OCS' not found
   added <- OCS$addGroup("scientists") #add a new group
#> Error: object 'OCS' not found
   sc_group <- OCS$getGroup("scientists") #get group details
#> Error: object 'OCS' not found
   added <- OCS$addToGroup("john.doe", "scientists") #add user to group
#> Error: object 'OCS' not found
   removed <- OCS$removeFromGroup("john.doe", "scientists") #remove user from group
#> Error: object 'OCS' not found
   deleted <- OCS$deleteGroup("scientists")
#> Error: object 'OCS' not found
   
   #OCS Webdav API
   #-------------------------------------
   #list files
   files <- OCS$listFiles()
#> Error: object 'OCS' not found
   subfiles <- OCS$listFiles("Documents")
#> Error: object 'OCS' not found
   #make collection
   OCS$makeCollection("myfolder")
#> Error: object 'OCS' not found
   subfiles <- OCS$listFiles("myfolder")
#> Error: object 'OCS' not found
   #upload a file?
   filename <- "magic.txt"
   file.create(filename); writeLines("ocs4R is great", filename)
#> [1] TRUE
   #we upload the file in 'Documents' folder
   OCS$uploadFile(filename, "/Documents")
#> Error: object 'OCS' not found
   #check if file is uploaded
   OCS$listFiles('Documents')
#> Error: object 'OCS' not found
   
   #OCS Sharing API
   #-------------------------------------
   #let's add a user with User provisioning API
   added <- OCS$addUser("john.doe", password = "ocs4john") #add a user
#> Error: object 'OCS' not found
   #let's share the previously uploaded file with John Doe
   OCS$shareWithUser("/Documents", filename, "john.doe")
#> Error: object 'OCS' not found
# }